requirements.txt
asgiref==3.7.2
build==1.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
Django==5.0.1
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
idna==3.6
packaging==25.0
pillow==10.2.0
pip-tools==7.5.0
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dotenv==1.0.1
requests==2.31.0
setuptools==80.9.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
wheel==0.45.1
